Added optimization for u8 arrays

+++ bytey_byte_buffer/src/byte_buffer_read/option.rs
@@ -7,7 +7,7 @@ use crate::{
 impl<T: ByteBufferRead> ByteBufferRead for Option<T> {
     #[inline]
     fn read_from_buffer(buffer: &mut ByteBuffer) -> Result<Option<T>> {
-        let data = match buffer.read::<u16>()? {
+        let data = match buffer.read::<u8>()? {
             1 => Some(buffer.read::<T>()?),
             2 => None,
             _ => {
@@ -22,7 +22,7 @@ impl<T: ByteBufferRead> ByteBufferRead for Option<T> {
 
     #[inline]
     fn read_from_buffer_le(buffer: &mut ByteBuffer) -> Result<Option<T>> {
-        let data = match buffer.read_le::<u16>()? {
+        let data = match buffer.read_le::<u8>()? {
             1 => Some(buffer.read_le::<T>()?),
             2 => None,
             _ => {
@@ -37,7 +37,7 @@ impl<T: ByteBufferRead> ByteBufferRead for Option<T> {
 
     #[inline]
     fn read_from_buffer_be(buffer: &mut ByteBuffer) -> Result<Option<T>> {
-        let data = match buffer.read_be::<u16>()? {
+        let data = match buffer.read_be::<u8>()? {
             1 => Some(buffer.read_be::<T>()?),
             2 => None,
             _ => {
