+++ bytey_byte_buffer/src/byte_buffer_read/result.rs
@@ -7,7 +7,7 @@ use crate::{
 impl<T: ByteBufferRead, E: ByteBufferRead> ByteBufferRead for std::result::Result<T, E> {
     #[inline]
     fn read_from_buffer(buffer: &mut ByteBuffer) -> Result<std::result::Result<T, E>> {
-        Ok(match buffer.read::<u16>()? {
+        Ok(match buffer.read::<u8>()? {
             1 => Ok(buffer.read::<T>()?),
             2 => Err(buffer.read::<E>()?),
             _ => {
@@ -20,7 +20,7 @@ impl<T: ByteBufferRead, E: ByteBufferRead> ByteBufferRead for std::result::Resul
 
     #[inline]
     fn read_from_buffer_le(buffer: &mut ByteBuffer) -> Result<std::result::Result<T, E>> {
-        Ok(match buffer.read_le::<u16>()? {
+        Ok(match buffer.read_le::<u8>()? {
             1 => Ok(buffer.read_le::<T>()?),
             2 => Err(buffer.read_le::<E>()?),
             _ => {
@@ -33,7 +33,7 @@ impl<T: ByteBufferRead, E: ByteBufferRead> ByteBufferRead for std::result::Resul
 
     #[inline]
     fn read_from_buffer_be(buffer: &mut ByteBuffer) -> Result<std::result::Result<T, E>> {
-        Ok(match buffer.read_be::<u16>()? {
+        Ok(match buffer.read_be::<u8>()? {
             1 => Ok(buffer.read_be::<T>()?),
             2 => Err(buffer.read_be::<E>()?),
             _ => {
@@ -48,7 +48,7 @@ impl<T: ByteBufferRead, E: ByteBufferRead> ByteBufferRead for std::result::Resul
 impl<E: ByteBufferRead> ByteBufferRead for std::result::Result<(), E> {
     #[inline]
     fn read_from_buffer(buffer: &mut ByteBuffer) -> Result<std::result::Result<(), E>> {
-        Ok(match buffer.read::<u16>()? {
+        Ok(match buffer.read::<u8>()? {
             1 => Ok(()),
             2 => Err(buffer.read::<E>()?),
             _ => {
@@ -61,7 +61,7 @@ impl<E: ByteBufferRead> ByteBufferRead for std::result::Result<(), E> {
 
     #[inline]
     fn read_from_buffer_le(buffer: &mut ByteBuffer) -> Result<std::result::Result<(), E>> {
-        Ok(match buffer.read_le::<u16>()? {
+        Ok(match buffer.read_le::<u8>()? {
             1 => Ok(()),
             2 => Err(buffer.read_le::<E>()?),
             _ => {
@@ -74,7 +74,7 @@ impl<E: ByteBufferRead> ByteBufferRead for std::result::Result<(), E> {
 
     #[inline]
     fn read_from_buffer_be(buffer: &mut ByteBuffer) -> Result<std::result::Result<(), E>> {
-        Ok(match buffer.read_be::<u16>()? {
+        Ok(match buffer.read_be::<u8>()? {
             1 => Ok(()),
             2 => Err(buffer.read_be::<E>()?),
             _ => {
